refactor(cards): move generic card components out from phone
