+++ src/components/phone/Widgets/CallHistory.vue
@@ -24,7 +24,7 @@
 </template>
 
 <script>
-import TitledCard from '@/components/phone/Cards/TitledCard.vue';
+import TitledCard from '@/components/cards/TitledCard.vue';
 import Table from '@/components/Table.vue';
 import { mapGetters } from 'vuex';
 import { UserMixin, ValidateMixin } from '@/mixins';
